meson: pass an header file when testing for posix_memalign
When compiling with MinGW, posix_memalign gets detected by
cc.has_function() but then compilation fails with the following error:

  /usr/bin/x86_64-w64-mingw32-ld: test/libtest-utils.a(utils.c.obj): in function `aligned_malloc':
  .../pixman/build/../test/utils.c:992: undefined reference to `posix_memalign'

This happens because the compiler has a builtin version of the function
but it does not expose it for linkage.

So pass an header file when checking for posix_memalign to make sure to
detect if the actual function is available.

This makes detection fail with MinGW and make the code compile.